task6

+++ hello_world/6-size.c
@@ -7,11 +7,11 @@
  */
 int main(void)
 {
-    printf("Size of char: %u bytes\n", (int) sizeof(char));
-    printf("Size of int: %u bytes\n", (int) sizeof(int));
-    printf("Size of long int: %u bytes\n", (int) sizeof(long int));
-    printf("Size of long long int: %u bytes\n", (int) sizeof(long long int));
-    printf("Size of float: %u bytes\n", (int) sizeof(float));
+    printf("Size of char: %lu bytes\n", sizeof(char));
+    printf("Size of int: %lu bytes\n",  sizeof(int));
+    printf("Size of long int: %lu bytes\n", sizeof(long int));
+    printf("Size of long long int: %lu bytes\n", sizeof(long long int));
+    printf("Size of float: %lu bytes\n", sizeof(float));
     return (0);
 }
